backend/flask/requirements.txt
flask==3.0.2
flask-cors==4.0.1
python-dotenv==1.0.1
pandas==2.2.2
langchain==0.3.1
google-generativeai==0.5.4
langchain-google-genai